@@ -15,6 +15,7 @@
     "engines": {
         "node": ">=12"
     },
+    "type": "module",
     "main": "index.js",
     "files": [
         "*.js"
@@ -35,7 +36,7 @@
         "@typescript-eslint/eslint-plugin": "^3.9.1",
         "@typescript-eslint/parser": "^3.9.1",
         "conventional-changelog-cli": "^2.1.0",
-        "eslint": "^7.7.0",
+        "eslint": "^7.8.1",
         "husky": "^4.2.5",
         "npm-run-all": "^4.1.5"
     },
