@@ -1,4 +1,7 @@
-exports.rules = {
-    'print-filename': require('./print-filename'),
-    'restrict-globals': require('./restrict-globals'),
+import * as printFileName from './print-filename.js';
+import * as restrictGlobals from './restrict-globals.js';
+
+export const rules = {
+    'print-filename': printFileName,
+    'restrict-globals': restrictGlobals,
 };
